Clang warnings solver (#1672)
Description
When compiling Lethe with Clang, multiple warning arise throughout the code. In the long term, we aim to remove these warnings so that Lethe compiles on both GCC and Clang without any warnings. All of the warnings of the solver library were removed. The majority of the warnings were related to some virtual classes not having a virtual distructors and some odd integers to double or double to integers conversions that were implicit. The type for the fill level in the ILU preconditioners was changed from double to integer since this is the actualy type required by the Trilinos library. This PR closes #1649.

Testing
All of the tests should pass without any issue. Nothing has changed there.

Documentation
This does not require documentation

+++ include/core/utilities.h
@@ -747,7 +747,7 @@ identify_minimum_cell_size(const Mapping<dim>    &mapping,
   double h = DBL_MAX;
 
   // Element degree
-  const double degree = double(fe_values.get_fe().degree);
+  const unsigned int degree = fe_values.get_fe().degree;
 
   for (const auto &cell : dof_handler.active_cell_iterators())
     {
